fix:  unwanted table component rerenders during scroll
